add support for AMD platforms
We can launch VMs on new platforms, but the performance monitor cannot work.
We use this flag to disable performance monitor for some platforms.

TODO: we need more changes to support AMD platforms as the PCM lib cannot
be used anymore and the performance counters are also different as Intel
platforms.